fix totals

+++ src/models.rs
@@ -102,7 +102,7 @@ pub mod db_models {
     pub struct Totals {
         pub bets_amount: i64,
         pub player_amount: i64,
-        pub sum: Option<f64>,
+        pub sum: Option<Decimal>,
     }
 
     // #[derive(Deserialize, Serialize, Clone, ToSchema)]
